Kategoriler eklendi.

+++ CafeAndRestaurantCheck_EF_Core/Forms/FrmKurulum.cs
@@ -109,6 +109,7 @@ namespace CafeAndRestaurantCheck_EF_Core.Forms
             var kategori = new Kategori();
             try
             {
+
                 kategori.Ad = txtKategoriAd.Text;
                 kategori.Aciklama = txtAciklama.Text;
 
@@ -164,12 +165,20 @@ namespace CafeAndRestaurantCheck_EF_Core.Forms
         }
         private void btnKategoriSil_Click(object sender, EventArgs e)
         {
-            var seciliKategori = (Kategori)this.dgViewKategori.CurrentRow.DataBoundItem;
+            var seciliKategori = new Kategori();
+            if (seciliKategori == null) return;
+            seciliKategori = (Kategori)this.dgViewKategori.CurrentRow.DataBoundItem;
             var kategori = _kategoriRepo.GetAll().FirstOrDefault(x => x.Id == seciliKategori.Id) as Kategori;
+            DialogResult cevap = MessageBox.Show($"{seciliKategori.Ad} yi silmek istiyor musunuz?", "Silme onayı", MessageBoxButtons.YesNo, MessageBoxIcon.Question);
+
+            if (cevap == DialogResult.Yes)
+            {
+                _kategoriRepo.Remove(kategori);
+                ListeleKategori();
+            }
             //var kategori = _kategoriRepo.GetById(seciliKategori.Id) as Kategori;
-            _kategoriRepo.Remove(kategori);
-            ListeleKategori();
-           
+
+
 
 
 
@@ -190,7 +199,9 @@ namespace CafeAndRestaurantCheck_EF_Core.Forms
 
         private void btnKategoriGuncelle_Click_1(object sender, EventArgs e)
         {
-            var seciliKategori = (Kategori)this.dgViewKategori.CurrentRow.DataBoundItem;
+            var seciliKategori = new Kategori();
+            if (seciliKategori == null) return;
+            seciliKategori = (Kategori)this.dgViewKategori.CurrentRow.DataBoundItem;
             var kategori = _kategoriRepo.GetById(seciliKategori.Id) as Kategori;
             seciliKategori.Ad = txtKategoriAd.Text;
             seciliKategori.Aciklama = txtAciklama.Text;
@@ -202,7 +213,20 @@ namespace CafeAndRestaurantCheck_EF_Core.Forms
 
         }
 
-        
+       
+
+        private void dgViewKategori_CellClick(object sender, DataGridViewCellEventArgs e)
+        {
+            seciliKategori = (Kategori)this.dgViewKategori.CurrentRow.DataBoundItem;
+            txtKategoriAd.Text = seciliKategori.Ad;
+            txtAciklama.Text = seciliKategori.Aciklama;
+            if (seciliKategori.Fotograf != null)
+            {
+                MemoryStream stream = new MemoryStream(seciliKategori.Fotograf);
+                pbKategori.Image = Image.FromStream(stream);
+
+            }
+        }
     }
 }
